+++ packages/apps/terminal/vite.config.ts
@@ -1,35 +1,6 @@
 import { defineConfig } from "vite";
-import { resolve } from "path";
-import react from "@vitejs/plugin-react-swc";
-import dts from "vite-plugin-dts";
-import { libInjectCss } from "vite-plugin-lib-inject-css";
+import { appViteConfig } from "@prozilla-os/shared";
 
-// https://vitejs.dev/config/
 export default defineConfig({
-	plugins: [
-		react(),
-		libInjectCss(),
-		dts({
-			include: ["src"],
-			outDir: "dist",
-			rollupTypes: true,
-			strictOutput: true,
-			pathsToAliases: false,
-			staticImport: true
-		})
-	],
-	build: {
-		lib: {
-			entry: resolve(__dirname, "src/main.ts"),
-			formats: ["es"],
-		},
-		rollupOptions: {
-			external: ["react", "react/jsx-runtime", "@prozilla-os/core"],
-			output: {
-				assetFileNames: "assets/[name][extname]",
-				entryFileNames: "[name].js",
-			}
-		},
-		sourcemap: true
-	}
+	...appViteConfig(__dirname, "src/main.ts")
 });