+++ packages/shared/src/configs/index.d.ts
@@ -1,0 +1,1 @@
+export { appViteConfig } from "./app.vite.config";