+++ packages/apps/terminal/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @prozilla-os/terminal
 
+## 1.0.4
+
+### Patch Changes
+
+- Added @prozilla-os/shared package
+  Updated app build process
+  Updated documentation
+- Updated dependencies
+  - @prozilla-os/core@1.0.14
+
 ## 1.0.3
 
 ### Patch Changes