+++ packages/prozilla-os/CHANGELOG.md
@@ -1,5 +1,18 @@
 # prozilla-os
 
+## 1.0.14
+
+### Patch Changes
+
+- Added @prozilla-os/shared package
+  Updated app build process
+  Updated documentation
+- Updated dependencies
+  - @prozilla-os/file-explorer@1.0.11
+  - @prozilla-os/text-editor@1.0.4
+  - @prozilla-os/terminal@1.0.4
+  - @prozilla-os/core@1.0.14
+
 ## 1.0.13
 
 ### Patch Changes